+++ packages/system/src/types/theme.ts
@@ -53,6 +53,7 @@ export type Theme = {
       'container' | 'icon',
       ComponentStyleFunction<string, string>
     >;
+    IconButton?: ComponentStyleFunction<string, string>;
     Image?: ComponentStyleFunction<string, string>;
     Checkbox?: Record<
       'container' | 'label' | 'checkbox' | 'group',
@@ -93,7 +94,11 @@ export type Theme = {
       | 'valueContainer',
       ComponentStyleFunction<string, string>
     >;
-    Pagination?: ComponentStyleFunction<string, string>;
+    Pagination?: Record<
+      'navigationButton' | 'pageButton' | 'icon',
+      ComponentStyleFunction<string, string>
+    >;
+    ProgressCycle?: ComponentStyleFunction<string, string>;
     Radio?: Record<
       'container' | 'label' | 'radio' | 'group',
       ComponentStyleFunction<string, string>
@@ -108,7 +113,7 @@ export type Theme = {
       ComponentStyleFunction<string, string>
     >;
     SectionMessage?: Record<
-      'container' | 'icon' | 'title' | 'content',
+      'container' | 'icon' | 'title' | 'content' | 'close',
       ComponentStyleFunction<string, string>
     >;
     Table?: Record<
@@ -138,10 +143,7 @@ export type Theme = {
       | 'calendarGrid',
       ComponentStyleFunction<string, string>
     >;
-    DatePicker?: Record<
-      'container' | 'button',
-      ComponentStyleFunction<string, string>
-    >;
+    DatePicker?: ComponentStyleFunction<string, string>;
     ComboBox?: ComponentStyleFunction<string, string>;
     XLoader?: Record<
       'container' | 'loader' | 'label',